backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.340
openai==0.28.0
chromadb==0.4.15
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2